Enum pot::Error

#[non_exhaustive]
pub enum Error {

    NotAPot,
    IncompatibleVersion,
    Message(String),
    TrailingBytes,
    Eof,
    ImpreciseCastWouldLoseData,
    Io(Error),
    SequenceSizeMustBeKnown,
    InvalidUtf8(String),
    InvalidKind(u8),
    UnexpectedKind(Kind, Kind),
    UnknownSymbol(u64),
    UnsupportedByteCount(Kind, usize),
    InvalidAtomHeader,
    TooManyBytesRead,
    UnknownSpecial(UnknownSpecial),

}

All errors that Pot may return.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

NotAPot

Payload is not a Pot payload.

IncompatibleVersion

Data was written with an incompatible version.

Message(String)

A generic error occurred.

TrailingBytes

Extra data appeared at the end of the input.

Eof

Expected more data but encountered the end of the input.

ImpreciseCastWouldLoseData

A numerical value could not be handled without losing precision or truncation.

Io(Error)

An IO error occurred.

SequenceSizeMustBeKnown

A sequence of unknown size cannot be serialized.

InvalidUtf8(String)

String data contained invalid UTF-8 characters.

InvalidKind(u8)

An unknown kind was encountered. Generally a sign that something else has been parsed incorrectly.

UnexpectedKind(Kind, Kind)

Encountered an unexpected atom kind.

UnknownSymbol(u64)

A requested symbol id was not found.

UnsupportedByteCount(Kind, usize)

An unsupported byte count for a numeric type was encountered.

InvalidAtomHeader

An atom header was incorrectly formatted.

TooManyBytesRead

The amount of data read exceeds the configured maximum number of bytes.

UnknownSpecial(UnknownSpecial)

An unknown Special was encountered.