Enum pot::format::Float

pub enum Float {
    F64(f64),
    F32(f32),
}

A floating point number that can safely convert between other number types using compile-time evaluation when possible.

Variants

F64(f64)

An f64 value.

F32(f32)

An f32 value.

Implementations

impl Float

pub fn as_f32(&self) -> Result<f32, Error>

Returns this number as an f32, if it can be done without losing precision.

#[must_use]
pub const fn as_f64(&self) -> f64

Returns this number as an f64.

pub fn as_integer(&self) -> Result<Integer, Error>

Returns this number as an Integer, if the stored value has no fractional part.

pub fn read_from<R: ReadBytesExt>(
kind: Kind,
byte_len: usize,
reader: &mut R
) -> Result<Self, Error>

Reads a floating point number given the atom kind and byte_len. byte_len should be the exact argument from the atom header.