Struct pot::ser::SymbolMap

pub struct SymbolMap { /* private fields */ }

A list of previously serialized symbols.

Implementations

impl SymbolMap

pub const fn new() -> Self

Returns a new, empty symbol map.

pub const fn with_compatibility(self, compatibility: Compatibility) -> Self

Sets the compatibility mode for serializing and returns self.

pub fn set_compatibility(&mut self, compatibility: Compatibility)

Sets the compatibility mode for serializing.

pub fn serializer_for<W: WriteBytesExt>( &mut self, output: W, ) -> Result<Serializer<'_, W>>

Returns a serializer that writes into output and persists symbols into self.

pub fn serialize_to<T, W>(&mut self, writer: W, value: &T) -> Result<()>
where W: Write, T: Serialize,

Serializes value into writer while persisting symbols into self.

pub fn serialize_to_vec<T>(&mut self, value: &T) -> Result<Vec<u8>>
where T: Serialize,

Serializes value into a new Vec<u8> while persisting symbols into self.

pub fn insert(&mut self, symbol: &str) -> bool

Inserts symbol into this map.

Returns true if this symbol had not previously been registered. Returns false if the symbol was already included in the map.

pub fn len(&self) -> usize

Returns the number of entries in this map.

pub fn is_empty(&self) -> bool

Returns true if the map has no entries.

pub fn populate_from<T>( &mut self, value: &T, ) -> Result<usize, SymbolMapPopulationError>
where T: Serialize,

Adds all symbols encountered in value.

Returns the number of symbols added.

Due to how serde works, this function can only encounter symbols that are being used. For example, if T is an enum, only variant being passed in will have its name, and additional calls for each variant will be needed to ensure every symbol is added.